// src/gql/graphql.ts
export type Maybe<T> = T | null
// ...
export type Scalars = {
	ID: { input: string; output: string }
	String: { input: string; output: string }
	Boolean: { input: boolean; output: boolean }
	Int: { input: number; output: number }
	Float: { input: number; output: number }
	Date: { input: any; output: any }
	ObjectID: { input: any; output: any }
	timestamptz: { input: any; output: any }
	uuid: { input: any; output: any }
}
// ...
export type Launch = {
	__typename?: 'Launch'
	details?: Maybe<Scalars['String']['output']>
	id?: Maybe<Scalars['ID']['output']>
	is_tentative?: Maybe<Scalars['Boolean']['output']>
	launch_date_local?: Maybe<Scalars['Date']['output']>
	launch_date_unix?: Maybe<Scalars['Date']['output']>
	launch_date_utc?: Maybe<Scalars['Date']['output']>
	launch_site?: Maybe<LaunchSite>
	launch_success?: Maybe<Scalars['Boolean']['output']>
	launch_year?: Maybe<Scalars['String']['output']>
	links?: Maybe<LaunchLinks>
	mission_id?: Maybe<Array<Maybe<Scalars['String']['output']>>>
	mission_name?: Maybe<Scalars['String']['output']>
	rocket?: Maybe<LaunchRocket>
	ships?: Maybe<Array<Maybe<Ship>>>
	static_fire_date_unix?: Maybe<Scalars['Date']['output']>
	static_fire_date_utc?: Maybe<Scalars['Date']['output']>
	telemetry?: Maybe<LaunchTelemetry>
	tentative_max_precision?: Maybe<Scalars['String']['output']>
	upcoming?: Maybe<Scalars['Boolean']['output']>
}
// ...
export type LaunchLinks = {
	__typename?: 'LaunchLinks'
	article_link?: Maybe<Scalars['String']['output']>
	flickr_images?: Maybe<Array<Maybe<Scalars['String']['output']>>>
	mission_patch?: Maybe<Scalars['String']['output']>
	mission_patch_small?: Maybe<Scalars['String']['output']>
	presskit?: Maybe<Scalars['String']['output']>
	reddit_campaign?: Maybe<Scalars['String']['output']>
	reddit_launch?: Maybe<Scalars['String']['output']>
	reddit_media?: Maybe<Scalars['String']['output']>
	reddit_recovery?: Maybe<Scalars['String']['output']>
	video_link?: Maybe<Scalars['String']['output']>
	wikipedia?: Maybe<Scalars['String']['output']>
}

export type LaunchRocket = {
	__typename?: 'LaunchRocket'
	fairings?: Maybe<LaunchRocketFairings>
	first_stage?: Maybe<LaunchRocketFirstStage>
	rocket?: Maybe<Rocket>
	rocket_name?: Maybe<Scalars['String']['output']>
	rocket_type?: Maybe<Scalars['String']['output']>
	second_stage?: Maybe<LaunchRocketSecondStage>
}

export type LaunchRocketFairings = {
	__typename?: 'LaunchRocketFairings'
	recovered?: Maybe<Scalars['Boolean']['output']>
	recovery_attempt?: Maybe<Scalars['Boolean']['output']>
	reused?: Maybe<Scalars['Boolean']['output']>
	ship?: Maybe<Scalars['String']['output']>
}

export type LaunchRocketFirstStage = {
	__typename?: 'LaunchRocketFirstStage'
	cores?: Maybe<Array<Maybe<LaunchRocketFirstStageCore>>>
}
// ...
export type LaunchRocketSecondStage = {
	__typename?: 'LaunchRocketSecondStage'
	block?: Maybe<Scalars['Int']['output']>
	payloads?: Maybe<Array<Maybe<Payload>>>
}

export type LaunchSite = {
	__typename?: 'LaunchSite'
	site_id?: Maybe<Scalars['String']['output']>
	site_name?: Maybe<Scalars['String']['output']>
	site_name_long?: Maybe<Scalars['String']['output']>
}

export type LaunchTelemetry = {
	__typename?: 'LaunchTelemetry'
	flight_club?: Maybe<Scalars['String']['output']>
}
// ...
export type Rocket = {
	__typename?: 'Rocket'
	active?: Maybe<Scalars['Boolean']['output']>
	boosters?: Maybe<Scalars['Int']['output']>
	company?: Maybe<Scalars['String']['output']>
	cost_per_launch?: Maybe<Scalars['Int']['output']>
	country?: Maybe<Scalars['String']['output']>
	description?: Maybe<Scalars['String']['output']>
	diameter?: Maybe<Distance>
	engines?: Maybe<RocketEngines>
	first_flight?: Maybe<Scalars['Date']['output']>
	first_stage?: Maybe<RocketFirstStage>
	height?: Maybe<Distance>
	id?: Maybe<Scalars['ID']['output']>
	landing_legs?: Maybe<RocketLandingLegs>
	mass?: Maybe<Mass>
	name?: Maybe<Scalars['String']['output']>
	payload_weights?: Maybe<Array<Maybe<RocketPayloadWeight>>>
	second_stage?: Maybe<RocketSecondStage>
	stages?: Maybe<Scalars['Int']['output']>
	success_rate_pct?: Maybe<Scalars['Int']['output']>
	type?: Maybe<Scalars['String']['output']>
	wikipedia?: Maybe<Scalars['String']['output']>
}
// ...
export type Ship = {
	__typename?: 'Ship'
	abs?: Maybe<Scalars['Int']['output']>
	active?: Maybe<Scalars['Boolean']['output']>
	attempted_landings?: Maybe<Scalars['Int']['output']>
	class?: Maybe<Scalars['Int']['output']>
	course_deg?: Maybe<Scalars['Int']['output']>
	home_port?: Maybe<Scalars['String']['output']>
	id?: Maybe<Scalars['ID']['output']>
	image?: Maybe<Scalars['String']['output']>
	imo?: Maybe<Scalars['Int']['output']>
	missions?: Maybe<Array<Maybe<ShipMission>>>
	mmsi?: Maybe<Scalars['Int']['output']>
	model?: Maybe<Scalars['String']['output']>
	name?: Maybe<Scalars['String']['output']>
	position?: Maybe<ShipLocation>
	roles?: Maybe<Array<Maybe<Scalars['String']['output']>>>
	speed_kn?: Maybe<Scalars['Float']['output']>
	status?: Maybe<Scalars['String']['output']>
	successful_landings?: Maybe<Scalars['Int']['output']>
	type?: Maybe<Scalars['String']['output']>
	url?: Maybe<Scalars['String']['output']>
	weight_kg?: Maybe<Scalars['Int']['output']>
	weight_lbs?: Maybe<Scalars['Int']['output']>
	year_built?: Maybe<Scalars['Int']['output']>
}

export type ShipLocation = {
	__typename?: 'ShipLocation'
	latitude?: Maybe<Scalars['Float']['output']>
	longitude?: Maybe<Scalars['Float']['output']>
}

export type ShipMission = {
	__typename?: 'ShipMission'
	flight?: Maybe<Scalars['String']['output']>
	name?: Maybe<Scalars['String']['output']>
}